refactor: LadderGame 패키지 수정
