+++ src/test/java/domain/model/LadderGameTest.java
@@ -1,13 +1,7 @@
-package domain.service;
+package domain.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.model.Ladder;
-import domain.model.Layer;
-import domain.model.Player;
-import domain.model.Players;
-import domain.model.Result;
-import domain.model.Results;
 import domain.vo.Height;
 import domain.vo.Width;
 import java.util.ArrayList;
